refractor: FE fileService -> cleanup indexeddb
